1.1.2 Equation improvements.
- Added functions abs, cos, sin, acos, asin, atan, sinh, sqrt, tan, tanh, truncate.
- Added equations (1.1)
- Fixed order of precedence faults so that symbols at the same level are evaluated left to right (e.g. addition and subtraction)
- Change equations (mult-line expressions) to use evaluate for rh values.
- Support scientific format numeric values (e.g. 1.2E+00)

An equation is an expression something like

var1=2 <LF>
var2=cos(var1)
=val+var1/var

Each line has an equals sign. If there is nothing on the left of the equals this is considered to be the return value. Otherwise a symbol will be set to the evaluated version of the RHS. External variables can be set (via eval.SetSymbol).

example of using curve fitting. External variables are iv.alpha and dv:
              <equation>
                a=-1.65798712337E-002
                b= 6.91771818666E-003
                c= 9.16997584955E-004
                d=-1.47185697480E-005
                e= 6.23910630630E-008
                x=iv.alpha
                =dv+(a+b*x+c*x^2+d*x^3+e*x^4)
              </equation>